Added redirect path support while login
